solutions/flask/requirements.txt
click==7.1.1
Flask==1.1.2
itsdangerous==1.1.0
more-itertools==8.2.0
packaging==20.3
py==1.8.1
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.1
python-dotenv==0.12.0
Werkzeug==1.0.1
